ISSUE-02: moved teller.py to package pythonsupermarket. adapted dependent files that use it.
