+++ pythonsupermarket/tests/test_supermarket.py
@@ -2,7 +2,7 @@ import pytest
 
 from pythonsupermarket.model_objects import Product, SpecialOfferType, ProductUnit
 from pythonsupermarket.shopping_cart import ShoppingCart
-from teller import Teller
+from pythonsupermarket.teller import Teller
 from tests.fake_catalog import FakeCatalog
